refactor: account -> member로 모두 수정
