+++ account-service/src/main/java/com/oopsw/memberservice/MemberServiceApplication.java
@@ -1,4 +1,4 @@
-package com.oopsw.accountservice;
+package com.oopsw.memberservice;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -8,7 +8,7 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 
 @SpringBootApplication
 @EnableDiscoveryClient
-public class AccountServiceApplication {
+public class MemberServiceApplication {
 
 	@Bean
 	public BCryptPasswordEncoder passwordEncoder() {
@@ -17,7 +17,7 @@ public class AccountServiceApplication {
 
 
 	public static void main(String[] args) {
-		SpringApplication.run(AccountServiceApplication.class, args);
+		SpringApplication.run(MemberServiceApplication.class, args);
 	}
 
 }
