+++ account-service/src/main/java/com/oopsw/memberservice/dto/MemberDto.java
@@ -1,11 +1,8 @@
-package com.oopsw.accountservice.dto;
+package com.oopsw.memberservice.dto;
 
 import java.util.Date;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 
 @Data
 public class MemberDto {
