+++ account-service/src/main/java/com/oopsw/memberservice/vo/request/ReqCheckNickname.java
@@ -1,4 +1,4 @@
-package com.oopsw.accountservice.vo.request;
+package com.oopsw.memberservice.vo.request;
 
 import lombok.Data;
