+++ account-service/src/main/java/com/oopsw/memberservice/vo/response/ResGetMember.java
@@ -1,4 +1,4 @@
-package com.oopsw.accountservice.vo.response;
+package com.oopsw.memberservice.vo.response;
 
 import java.util.Date;
